+++ src/components/ui/toolbox.ts
@@ -3,9 +3,9 @@ import { BlockToolAPI } from '../block';
 import Shortcuts from '../utils/shortcuts';
 import BlockTool from '../tools/block';
 import ToolsCollection from '../tools/collection';
-import { API } from '../../../types';
+import { API, BlockToolData, ToolboxConfigEntry } from '../../../types';
 import EventsDispatcher from '../utils/events';
-import Popover, { PopoverEvent } from '../utils/popover';
+import Popover, { PopoverEvent, PopoverItem } from '../utils/popover';
 import I18n from '../i18n';
 import { I18nInternalNS } from '../i18n/namespace-internal';
 
@@ -132,17 +132,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
       searchable: true,
       filterLabel: this.i18nLabels.filter,
       nothingFoundLabel: this.i18nLabels.nothingFound,
-      items: this.toolsToBeDisplayed.map(tool => {
-        return {
-          icon: tool.toolbox.icon,
-          label: I18n.t(I18nInternalNS.toolNames, tool.toolbox.title || _.capitalize(tool.name)),
-          name: tool.name,
-          onClick: (item): void => {
-            this.toolButtonActivated(tool.name);
-          },
-          secondaryLabel: tool.shortcut ? _.beautifyShortcut(tool.shortcut) : '',
-        };
-      }),
+      items: this.toolboxItemsToBeDisplayed,
     });
 
     this.popover.on(PopoverEvent.OverlayClicked, this.onOverlayClicked);
@@ -185,9 +175,10 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
    * Toolbox Tool's button click handler
    *
    * @param toolName - tool type to be activated
+   * @param blockDataOverrides - Block data predefined by the activated Toolbox item
    */
-  public toolButtonActivated(toolName: string): void {
-    this.insertNewBlock(toolName);
+  public toolButtonActivated(toolName: string, blockDataOverrides: BlockToolData): void {
+    this.insertNewBlock(toolName, blockDataOverrides);
   }
 
   /**
@@ -262,24 +253,79 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
   private get toolsToBeDisplayed(): BlockTool[] {
     return Array
       .from(this.tools.values())
-      .filter(tool => {
+      .reduce((result, tool) => {
         const toolToolboxSettings = tool.toolbox;
 
-        /**
-         * Skip tools that don't pass 'toolbox' property
-         */
-        if (!toolToolboxSettings) {
-          return false;
+        if (toolToolboxSettings) {
+          const validToolboxSettings = toolToolboxSettings.filter(item => {
+            return this.areToolboxSettingsValid(item, tool.name);
+          });
+
+          result.push({
+            ...tool,
+            toolbox: validToolboxSettings,
+          });
         }
 
-        if (toolToolboxSettings && !toolToolboxSettings.icon) {
-          _.log('Toolbar icon is missed. Tool %o skipped', 'warn', tool.name);
+        return result;
+      }, []);
+  }
 
-          return false;
+  /**
+   * Returns list of items that will be displayed in toolbox
+   */
+  @_.cacheable
+  private get toolboxItemsToBeDisplayed(): PopoverItem[] {
+    /**
+     * Maps tool data to popover item structure
+     */
+    const toPopoverItem = (toolboxItem: ToolboxConfigEntry, tool: BlockTool): PopoverItem => {
+      return {
+        icon: toolboxItem.icon,
+        label: I18n.t(I18nInternalNS.toolNames, toolboxItem.title || _.capitalize(tool.name)),
+        name: tool.name,
+        onClick: (e): void => {
+          this.toolButtonActivated(tool.name, toolboxItem.data);
+        },
+        secondaryLabel: tool.shortcut ? _.beautifyShortcut(tool.shortcut) : '',
+      };
+    };
+
+    return this.toolsToBeDisplayed
+      .reduce((result, tool) => {
+        if (Array.isArray(tool.toolbox)) {
+          tool.toolbox.forEach(item => {
+            result.push(toPopoverItem(item, tool));
+          });
+        } else {
+          result.push(toPopoverItem(tool.toolbox, tool));
         }
 
-        return true;
-      });
+        return result;
+      }, []);
+  }
+
+  /**
+   * Validates tool's toolbox settings
+   *
+   * @param toolToolboxSettings - item to validate
+   * @param toolName - name of the tool used in console warning if item is not valid
+   */
+  private areToolboxSettingsValid(toolToolboxSettings: ToolboxConfigEntry, toolName: string): boolean {
+    /**
+     * Skip tools that don't pass 'toolbox' property
+     */
+    if (!toolToolboxSettings) {
+      return false;
+    }
+
+    if (toolToolboxSettings && !toolToolboxSettings.icon) {
+      _.log('Toolbar icon is missed. Tool %o skipped', 'warn', toolName);
+
+      return false;
+    }
+
+    return true;
   }
 
   /**
@@ -331,8 +377,9 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
    * Can be called when button clicked on Toolbox or by ShortcutData
    *
    * @param {string} toolName - Tool name
+   * @param blockDataOverrides - predefined Block data
    */
-  private insertNewBlock(toolName: string): void {
+  private async insertNewBlock(toolName: string, blockDataOverrides?: BlockToolData): Promise<void> {
     const currentBlockIndex = this.api.blocks.getCurrentBlockIndex();
     const currentBlock = this.api.blocks.getBlockByIndex(currentBlockIndex);
 
@@ -346,9 +393,20 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
      */
     const index = currentBlock.isEmpty ? currentBlockIndex : currentBlockIndex + 1;
 
+    let blockData;
+
+    if (blockDataOverrides) {
+      /**
+       * Merge real tool's data with data overrides
+       */
+      const defaultBlockData = await this.api.blocks.composeBlockData(toolName);
+
+      blockData = Object.assign(defaultBlockData, blockDataOverrides);
+    }
+
     const newBlock = this.api.blocks.insert(
       toolName,
-      undefined,
+      blockData,
       undefined,
       index,
       undefined,