+++ docs/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 2.24.4
 
-- `Fix` - `Paste` - Replace the whole content of editor with pasted data.[#1705](https://github.com/codex-team/editor.js/issues/1705)
+- `Fix` — Keyboard selection by word [2045](https://github.com/codex-team/editor.js/issues/2045)
+- `Fix` - `Paste` - Replace the whole content of editor with pasted data.[#1705](https://github.com/codex-team/editor.js/issues/1705).
 
 ### 2.24.3